requirement.txt
Flask==3.0.3
Flask-Cors==4.0.0
python-dotenv==1.0.1
easyocr==1.7.1
pdf2image==1.17.0
google-generativeai==0.7.2
Pillow==10.3.0
Werkzeug==3.0.3
pyspellchecker==0.8.1
torch==2.2.2
torchvision==0.17.2
tqdm==4.66.5
numpy==1.26.4
opencv-python-headless==4.9.0.80